Automatically Create a shiny App Based on Interactive Loon Widgets

Interactive loon widgets displayed in shiny app

shiny.loon(
  widgets,
  showWorldView = TRUE,
  plotWidth = "500px",
  plotHeight = "500px",
  selectBy = c("byDefault", "brushing", "sweeping"),
  layout_matrix,
  nrow = NULL,
  ncol = NULL,
  colorList = loon::l_getColorList(),
  top = 60,
  left = "auto",
  right = 20,
  bottom = "auto",
  inspectorWidth = 350,
  inspectorHeight = "auto",
  options = list(),
  arrangeGrobArgs = list(),
  ...
)

Arguments

widgets
a loon widget or a list of loon widgets.

showWorldView
logical; whether to show the world view.

plotWidth
Image width, same with height.

plotHeight
Image height. Must be a valid CSS unit (like '100 which will be coerced to a string and have 'px' appended. More details, see plotOutput.

selectBy
The way to brush, can be 'brushing' (keep the brush whenever the plot is updated), 'sweeping' (clear the brush whenever the plot is updated) or 'byDefault' (determined by loon widget 'selectBy', if 'selectBy' is sweeping, then selectBy will be assigned to 'sweeping', vice versa)

layout_matrix
Optional layout matrix to place loon widgets, see grid.arrange.

nrow
Number of rows, see grid.arrange.

ncol
Number of columns, see grid.arrange.

colorList
A list of colors displayed on modify panel.

top
Distance between the top of the inspector panel, and the top of the page or parent container.

left
Distance between the left side of the inspector panel, and the left of the page or parent container.

right
Distance between the right side of the inspector panel, and the right of the page or parent container.

bottom
Distance between the bottom of the inspector panel, and the bottom of the page or parent container.

inspectorWidth
Width of the inspector panel.

inspectorHeight
Height of the inspector panel.

options
shinyApp argument that should be passed to the runApp call, see shinyApp.

arrangeGrobArgs
Other arrangeGrob arguments to place the loon widgets, see grid.arrange.

...
Named arguments to modify shiny app.

Value

A shiny.appobj object that represents the loon.shiny app. Printing the object or passing it to runApp will run the app.

Details

Useful hints:

The effect of any actionButtons can be terminated by a simple click on canvas.

The choices of 'select_by_color' can be updated by double click on color 'apply' actionButton.

Examples

## Only run this example in interactive R sessions
if(interactive()) {
  ############### Basic ###############
  p <- l_plot(iris, 
              color = iris$Species, 
              showGuides = TRUE, 
              showScales = TRUE)

  shiny.loon(p)
  
  ############### Link multiple plots ###############
  p1 <- l_plot(iris,
               linkingGroup = 'iris', 
               showLabels = FALSE)
  p2 <- l_hist(iris$Sepal.Length, 
               linkingGroup = 'iris', 
               showLabels = FALSE, 
               showStackedColors = TRUE)
  p3 <- l_hist(iris$Sepal.Width, linkingGroup = 'iris', 
               color = iris$Species, sync = 'push',
               showLabels = FALSE, swapAxes = TRUE, 
               showStackedColors = TRUE)
  shiny.loon(list(p1, p2, p3), 
             layout_matrix = matrix(c(2,NA,1,3), 
             nrow = 2, byrow = TRUE))
}

if (FALSE) {
  if (requireNamespace('loon.ggplot', quietly = TRUE)) {
      p <- ggplot(mpg, aes(displ, hwy)) +
        geom_point(data = transform(mpg, class = NULL), colour = 'grey85') +
        geom_point() +
        facet_wrap(~class)
      g <- loon.ggplot(p, activeGeomLayers = 2) # active the second layer
      shiny.loon(g)
  }
}