shiny.loon (deprecated)

Package loon.shiny only provides one exported function. To make it easier for users to remember, we decide to name the function with the selfsame name, loon.shiny.

shiny.loon(
  widgets,
  selectBy = c("byDefault", "brushing", "sweeping"),
  showWorldView = TRUE,
  plotRegionWidth = "500px",
  plotRegionHeight = "500px",
  plotRegionBackground = "gray92",
  layoutMatrix = NULL,
  nrow = NULL,
  ncol = NULL,
  widths = NULL,
  heights = NULL,
  displayedPanel = "",
  colorList = loon::l_getColorList(),
  inspectorLocation = c("auto", "auto", "60px", "20px"),
  inspectorWidth = "350px",
  inspectorHeight = "auto",
  options = list(),
  ...
)

Arguments

widgets

A loon widget or a list of loon widgets. If the input is a ggplot object, the ggplot object will be turned into a loon widget via ggplot2loon.

selectBy

The way to brush, can be 'brushing' (keep the brush whenever the plot is updated), 'sweeping' (clear the brush whenever the plot is updated) or 'byDefault' (determined by loon widget 'selectBy')

showWorldView

Logical; whether to show the world view.

plotRegionWidth

Plot region width. Must be a valid CSS unit (like '100 which will be coerced to a string and have 'px' appended.

plotRegionHeight

Plot region height.

plotRegionBackground

Plot region background color

layoutMatrix

Optional layout matrix to place loon widgets. See layout_matrix in grid.arrange.

nrow

Number of rows, see grid.arrange.

ncol

Number of columns, see grid.arrange.

widths

A unit vector giving the width of each plot.

heights

A unit vector giving the height of each plot.

displayedPanel

A string vector. The default is an empty string "" so that none inspector components (Plot, Linking, Select, etc) are open automatically. The available strings are c("Plot", "Select", "Linking", "Modify", "Layer", "Glyph")

colorList

A list of colors displayed on modify panel.

inspectorLocation

A length four vector representing the distance between the bottom, left, top and right of the inspector panel and the bottom, left, top and right of the page or parent container.

inspectorWidth

Width of the inspector panel.

inspectorHeight

Height of the inspector panel.

options

shinyApp argument that should be passed to the runApp call, see shinyApp.

...

Named arguments to modify shiny app.